Converted snake-game to a Meteor project - Current front-end is Blaze - Everything implemented in main.js for the sake of "please-please-please-let-this-work". - Added an END button to stop the game before dying.
