+++ .meteor/release
@@ -1,0 +1,1 @@
+METEOR@1.6.1
