+++ .meteor/versions
@@ -1,0 +1,78 @@
+allow-deny@1.1.0
+autopublish@1.0.7
+autoupdate@1.4.0
+babel-compiler@7.0.0
+babel-runtime@1.2.2
+base64@1.0.10
+binary-heap@1.0.10
+blaze@2.3.2
+blaze-html-templates@1.1.2
+blaze-tools@1.0.10
+boilerplate-generator@1.4.0
+caching-compiler@1.1.11
+caching-html-compiler@1.1.2
+callback-hook@1.1.0
+check@1.3.0
+ddp@1.4.0
+ddp-client@2.3.1
+ddp-common@1.4.0
+ddp-server@2.1.2
+deps@1.0.12
+diff-sequence@1.1.0
+dynamic-import@0.3.0
+ecmascript@0.10.0
+ecmascript-runtime@0.5.0
+ecmascript-runtime-client@0.6.0
+ecmascript-runtime-server@0.5.0
+ejson@1.1.0
+es5-shim@4.7.3
+geojson-utils@1.0.10
+hot-code-push@1.0.4
+html-tools@1.0.11
+htmljs@1.0.11
+http@1.4.0
+id-map@1.1.0
+insecure@1.0.7
+jquery@1.11.10
+launch-screen@1.1.1
+livedata@1.0.18
+logging@1.1.19
+meteor@1.8.2
+meteor-base@1.3.0
+minifier-css@1.3.0
+minifier-js@2.3.1
+minimongo@1.4.3
+mobile-experience@1.0.5
+mobile-status-bar@1.0.14
+modules@0.11.3
+modules-runtime@0.9.2
+mongo@1.4.2
+mongo-dev-server@1.1.0
+mongo-id@1.0.6
+npm-mongo@2.2.34
+observe-sequence@1.0.16
+ordered-dict@1.1.0
+promise@0.10.1
+random@1.1.0
+reactive-var@1.0.11
+reload@1.2.0
+retry@1.1.0
+routepolicy@1.0.12
+server-render@0.3.0
+shell-server@0.3.1
+shim-common@0.1.0
+socket-stream-client@0.1.0
+spacebars@1.0.15
+spacebars-compiler@1.1.3
+standard-minifier-css@1.4.0
+standard-minifier-js@2.3.1
+templating@1.3.2
+templating-compiler@1.3.3
+templating-runtime@1.3.2
+templating-tools@1.1.2
+tracker@1.1.3
+ui@1.0.13
+underscore@1.0.10
+url@1.2.0
+webapp@1.5.0
+webapp-hashing@1.0.9
