+++ client/main.css
@@ -1,0 +1,1 @@
+/* CSS declarations go here */
