+++ client/main.html
@@ -1,0 +1,22 @@
+<head>
+  <title>snake</title>
+</head>
+
+<body>
+<h1>Welcome to Snake!</h1>
+
+{{> game}}
+</body>
+
+<template name="game">
+
+  <canvas id="canvas" width = "700" height = "600">
+  </canvas>
+
+  <p>
+    Press start and eat the pizza!
+    <button id='btn'>START</button>
+    <button id='btn-end'>EXIT</button>
+  </p>
+
+</template>
