--- js/setting.js
@@ -1,10 +1,0 @@
-
-var mycanvas = document.getElementById('mycanvas');
-var ctx = mycanvas.getContext('2d');
-var snakeSize = 10; 
-var w = 350;
-var h = 350;
-var score = 0;
-var snake;
-var snakeSize = 10;
-var food;
