--- snakegame.html
@@ -1,22 +1,0 @@
-<html>
- <head>
-  <meta charset='utf-8'/>
-  	<link rel="stylesheet" href="css/style.css">
-
- </head>
- <body>
-  <div class= 'game'>
-  <div id = 'home'>
-    <canvas id='mycanvas' width='350' height='350'>
-    </canvas>
-    </div>
-    <p>Press start and eat the pizza!</p>
-    <button id='btn'>START</button>
-  </div>
-
-  <script src="js/setting.js"></script>
-  <script src="js/draw.js"></script>
-  <script src="js/app.js"></script>
- </body>
-
-</html>
